+++ src/components/explore-section/EphysViewer/components/StimulusPlot.tsx
