+++ src/components/explore-section/EphysViewer/components/TraceOverview.tsx
@@ -2,8 +2,9 @@ import { useMemo, useState } from 'react';
 import Plotly from 'plotly.js-dist-min';
 import { Select } from 'antd';
 import { LineChartOutlined } from '@ant-design/icons';
+import startCase from 'lodash/startCase';
 
-import NWBTrace from './nwb-trace';
+import NWBTrace from '../nwb-trace';
 import { useInView } from 'react-intersection-observer';
 import createPlotlyComponent from 'react-plotly.js/factory';
 import optimizePlotData from '@/util/explore-section/optimizeTrace';
@@ -47,7 +48,7 @@ function TraceThumbnail({
 
   const sweeps = trace.getSweeps(protocol, repetition);
 
-  const rawData = useMemo(() => {
+  const [rawData, dataUnit] = useMemo(() => {
     let deltaTime = 1;
     let dataUnit: string | null = null;
     let conversionFactor = 1;
@@ -60,7 +61,6 @@ function TraceThumbnail({
 
         if (timeUnit === 'seconds') {
           deltaTime = (1 / timeRate) * 1000;
-          console.log('deltaTime', deltaTime);
         }
 
         dataUnit = sweepData[recordingType].unit;
@@ -95,8 +95,10 @@ function TraceThumbnail({
           : convertVoltageSeries(d.y, 'mV', conversionFactor);
     });
 
-    return optimizedPlotData;
+    return [optimizedPlotData, dataUnit];
   }, [sweeps]);
+
+  const yTitle = `${startCase(recordingType)} (${dataUnit === 'amperes' ? 'pA' : 'mV'})`;
 
   return (
     <Plot
@@ -105,23 +107,59 @@ function TraceThumbnail({
       onDoubleClick={() => false}
       style={{ width: '100%', height: '100%' }}
       layout={{
+        shapes: [
+          {
+            type: 'rect',
+            xref: 'paper',
+            yref: 'paper',
+            x0: 0,
+            y0: 0,
+            x1: 1,
+            y1: 1,
+            line: {
+              color: '#808080',
+              width: 1,
+            },
+          },
+        ],
         showlegend: false,
         font: {
           size: 10,
         },
         margin: {
-          l: 20,
+          l: 48,
           r: 0,
           t: 0,
-          b: 20,
+          b: 42,
         },
         xaxis: {
+          ticks: 'outside',
+          ticklen: 6,
+          tickwidth: 1,
+          tickcolor: 'black',
+          automargin: true,
           zeroline: false,
+          title: {
+            font: {
+              size: 12,
+            },
+            text: 'Time (ms)',
+          },
         },
         yaxis: {
+          ticks: 'outside',
+          ticklen: 4,
+          tickwidth: 1,
+          tickcolor: 'black',
+          automargin: true,
           zeroline: false,
+          title: {
+            font: {
+              size: 12,
+            },
+            text: yTitle,
+          },
         },
-        // autosize: true,
       }}
       config={{ displaylogo: false, responsive: true, staticPlot: true }}
     />
@@ -139,10 +177,10 @@ function TraceThumbnailContainer({
   repetition: string;
   recordingType: 'stimulus' | 'response';
 }) {
-  const [ref, inView] = useInView({ threshold: 0, triggerOnce: true, rootMargin: '1200px 0px' });
+  const [ref, inView] = useInView({ threshold: 0, triggerOnce: true, rootMargin: '1200px' });
 
   return (
-    <div ref={ref} className="relative aspect-video w-full border border-lime-400">
+    <div ref={ref} className="aspect-4/3 relative w-full overflow-hidden last:mt-7">
       {inView ? (
         <TraceThumbnail
           trace={trace}
@@ -190,17 +228,13 @@ function ImageSetComponent({
             </div>
 
             {['stimulus', 'response'].map((recordingType: string) => (
-              <div className="flex items-center" key={recordingType}>
-                <span className="-rotate-90 capitalize text-neutral-4">{recordingType}</span>
-                {/* {imagePreview({ imageUrl: imgData.imageSrc })} */}
-                <TraceThumbnailContainer
-                  trace={trace}
-                  protocol={protocol}
-                  repetition={repetition}
-                  recordingType={recordingType}
-                />
-                {/* <div className="h-32">test</div> */}
-              </div>
+              <TraceThumbnailContainer
+                key={recordingType}
+                trace={trace}
+                protocol={protocol}
+                repetition={repetition}
+                recordingType={recordingType}
+              />
             ))}
           </div>
         ))}
