+++ src/components/explore-section/EphysViewer/index.tsx
@@ -1,81 +1,5 @@
-import { useState } from 'react';
-import { Empty, Radio, RadioChangeEvent, Spin } from 'antd';
-import { FileImageOutlined, LineChartOutlined } from '@ant-design/icons';
-import { useAtomValue } from 'jotai';
+import dynamic from 'next/dynamic';
 
-import { ExperimentalTrace } from '@/types/explore-section/delta-experiment';
-import TraceOverview from '@/components/explore-section/EphysViewer/TraceOverview';
-import TraceDetailsView from './TraceDetailsView';
-import sessionAtom from '@/state/session';
-import useTrace from './hooks/use-nwb-trace';
+const EphysViewer = dynamic(() => import('./EphysViewer'), { ssr: false });
 
-import './styles/ephys-plugin-styles.scss';
-
-enum VIEW {
-  OVERVIEW = 'overview',
-  DETAILED = 'detailed',
-}
-
-function EphysViewerContainer({ resource }: { resource: ExperimentalTrace }) {
-  const session = useAtomValue(sessionAtom);
-  const [trace, error] = useTrace(resource, session);
-
-  const [view, setView] = useState<VIEW>(VIEW.OVERVIEW);
-  const [repetition, setRepetition] = useState<string>();
-  const [protocol, setProtocol] = useState<string>('All');
-
-  const handleViewChange = (e: RadioChangeEvent) => {
-    setView(e.target.value as VIEW);
-  };
-
-  const showRepetitionDetails = (protocol: string, repetition: string) => () => {
-    setProtocol(protocol);
-    setRepetition(repetition);
-    setView(VIEW.DETAILED);
-  };
-
-  if (error) {
-    console.error(error);
-
-    return (
-      <Empty className="p-2em" description="There was a problem loading the required resources" />
-    );
-  }
-
-  if (!trace) {
-    return <Spin />;
-  }
-
-  return (
-    <div className="flex flex-col gap-6">
-      <Radio.Group onChange={handleViewChange} value={view}>
-        <Radio.Button value={VIEW.OVERVIEW}>
-          <FileImageOutlined /> Overview
-        </Radio.Button>
-
-        <Radio.Button value={VIEW.DETAILED}>
-          <LineChartOutlined /> Interactive Details
-        </Radio.Button>
-      </Radio.Group>
-
-      {view === VIEW.OVERVIEW && (
-        <TraceOverview
-          trace={trace}
-          stimulusType={protocol}
-          onRepetitionClick={showRepetitionDetails}
-          onStimulusChange={setProtocol}
-        />
-      )}
-
-      {view === VIEW.DETAILED && (
-        <TraceDetailsView
-          trace={trace}
-          defaultProtocol={protocol === 'None' || protocol === 'All' ? undefined : protocol}
-          defaultRepetition={repetition}
-        />
-      )}
-    </div>
-  );
-}
-
-export default EphysViewerContainer;
+export default EphysViewer;
